ifGranted () used | Datatable added
ifgranted() function used to limit access to different parts of the application
for example the user used to be able to make new entries of everything to the database. Now it can only make new entries for recepten. Another example, the apotheker cannot make any new entries into the database, it doesn't have access and if it tries to get into the page by hacking or whatever it will be denied.

I have also used the MDBootstrap datatable extension for the sorting abilities it offers me and a nice extra is the pagination which was given with it. The pagination ensures only 10 entries can show up on one page but it can be increased by clicking on a button so it works pretty nicely.